refactor(parser): split `parser.rs` into multiple files
