+++ javac-rs-parser/src/parser/tests/mod.rs
@@ -1,0 +1,3 @@
+use super::*;
+
+mod expression;
